requirements.txt
pillow >= 12.0.0
pymupdf >= 1.26.5

# For testing
pytest >= 8.4.2

# For building
pyinstaller >= 6.16.0